// src/pages/CostApplyManage/typings.d.ts
// @ts-ignore
/* eslint-disable */

declare namespace API {
  type Params = {
    applyStatus: number,
    beginTime: string,
    endTime: string,
    pageNumber: number,
    pageSize: number,
    userName: string
  };

  type TableItem = {
    applyAmount: number,
    applyStatus: number,
    applyTime: string,
    costApplyId: number,
    costApplyNo: string,
    remark: string,
    userId: number,
    userName: string
  }
}